fix: resolve remaining MSVC warnings (int→float, Uint64→uint32_t)

+++ src/draw.c
@@ -4,17 +4,17 @@
 void bapi_engine_render_drawpixel(int x, int y, int color)
 {
 	bapi_color_t c = bapi_color_from_hex((uint32_t)color);
-	bapi_draw_pixel(x, y, c);
+	bapi_draw_pixel((float)x, (float)y, c);
 }
 
 void bapi_engine_render_fillrect(int ax, int ay, int width, int height, int color)
 {
 	bapi_color_t c = bapi_color_from_hex((uint32_t)color);
-	bapi_fill_rect(ax, ay, width, height, c);
+	bapi_fill_rect((float)ax, (float)ay, (float)width, (float)height, c);
 }
 
 void bapi_engine_render_draw_triangle(int x1, int y1, int x2, int y2, int x3, int y3, int color)
 {
 	bapi_color_t c = bapi_color_from_hex((uint32_t)color);
-	bapi_draw_triangle(x1, y1, x2, y2, x3, y3, c);
+	bapi_draw_triangle((float)x1, (float)y1, (float)x2, (float)y2, (float)x3, (float)y3, c);
 }
